I've settled down, but...
neovim has many bugs, such as ]f or [f  that crash the app completely. It's implementation is bad due to the rotten vim code that preceeded it. That being said, this is probably the last changes that I'll make in it, because it is not worth it anymore to configure it, especially if it is going to break down the road.

I'll wait for it to become more stable.